Adds option to skip develop merge
Introduces an `onlyMaster` option to hotfix and release flows.

This allows skipping the merge into the develop branch, providing more
flexibility in scenarios where changes should only be applied to master, like after the user resolves a conflict while merging into the develop branch and the finish git flow continues but it only needs to be applied to master.